# SWAGGER_API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: Lilium B2B E-commerce Backend API
  description: Complete API documentation for the Lilium B2B multi-vendor platform with cash-on-delivery system
  version: 1.0.0
  contact:
    name: Lilium API Support
    email: support@lilium.iq
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.lilium.iq/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        parentId:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List all categories
      parameters:
        - name: tree
          in: query
          schema:
            type: boolean
            description: Return as tree structure
      responses:
        200:
          description: List of categories

    post:
      tags: [Categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nameAr, nameEn]
              properties:
                nameAr:
                  type: string
                nameEn:
                  type: string
                parentId:
                  type: string
                sortOrder:
                  type: integer
      responses:
        201:
          description: Category created
  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get category details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Category details

    put:
      tags: [Categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          description: Category updated

    delete:
      tags: [Categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: reassignTo
          in: query
          schema:
            type: string
            description: Category ID to reassign products to
      responses:
        200:
          description: Category deleted
...
  /categories/reorder:
    patch:
      tags: [Categories]
      summary: Reorder categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  sortOrder:
                    type: integer
      responses:
        200:
          description: Categories reordered

  # ==================== ORDERS ====================
...